ACS-306: ACS Repository: V1 Rest API for content direct access urls    - change expiryTime type to long    - updated alfresco data model version

+++ src/main/java/org/alfresco/repo/content/ContentServiceImpl.java
@@ -509,7 +509,7 @@ public class ContentServiceImpl extends ContentTransformServiceAdaptor implement
     }
 
     @Override
-    public String getDirectAccessUrl(NodeRef nodeRef, int expiryTime)
+    public String getDirectAccessUrl(NodeRef nodeRef, long expiryTime)
     {
         ContentData contentData = getContentData(nodeRef, ContentModel.PROP_CONTENT);
